CORE-1356 Move user info mocks to their own file
