+++ stories/data/DataMocks.js
@@ -495,87 +495,6 @@ export const plainChunk = {
         '"It was a dark and stormy night; the rain fell in torrents--except at occasional intervals, when it was checked by a violent gust of wind which swept up the streets (for it is in London that our scene lies), rattling along the housetops, and fiercely agitating the scanty flame of the lamps that struggled against the darkness."\n',
 };
 
-export const GROUP_ID = "1448e0f77d794bc9965dea4bf8eddecb";
-
-export const userInfoResp = {
-    batman: {
-        email: "batman@iplantcollaborative.org",
-        name: "Bruce Wayne",
-        last_name: "Wayne",
-        description: "He is the night",
-        id: "batman",
-        institution: "The Night",
-        first_name: "Bruce",
-        source_id: "ldap",
-    },
-    robin: {
-        email: "robin@cyverse.org",
-        name: "Richard Grayson",
-        last_name: "Grayson",
-        description: "Sidekick",
-        id: "robin",
-        institution: "The Night",
-        first_name: "Richard",
-        source_id: "ldap",
-    },
-    alfred: {
-        email: "alfred@cyverse.org",
-        name: "Alfred Pennyworth",
-        last_name: "Pennyworth",
-        description: "Unappreciated",
-        id: "alfred",
-        institution: "Bat Cave",
-        first_name: "Alfred",
-        source_id: "ldap",
-    },
-    catwoman: {
-        email: "catwoman@cyverse.org",
-        name: "Selina Kyle",
-        last_name: "Kyle",
-        description: "Sneaky",
-        id: "catwoman",
-        institution: "Catnip",
-        first_name: "Selina",
-        source_id: "ldap",
-    },
-    joker: {
-        email: "joker@cyverse.org",
-        name: "Jack Napier",
-        last_name: "Napier",
-        description: "Insane",
-        id: "joker",
-        institution: "Madness",
-        first_name: "Jack",
-        source_id: "ldap",
-    },
-    penguin: {
-        email: "penguin@cyverse.org",
-        name: "Oswald Cobblepot",
-        last_name: "Cobblepot",
-        description: "Waddler",
-        id: "penguin",
-        institution: "Penguins Inc",
-        first_name: "Oswald",
-        source_id: "ldap",
-    },
-    freeze: {
-        email: "freeze@cyverse.org",
-        name: "Victor Fries",
-        last_name: "Fries",
-        description: "Cold",
-        id: "freeze",
-        institution: "Ice To Meet You",
-        first_name: "Victor",
-        source_id: "ldap",
-    },
-    [GROUP_ID]: {
-        name: "iplant:de:qa:teams:freeze:gotham_gang",
-        description: "Gotham Baddies",
-        id: GROUP_ID,
-        source_id: "g:gsa",
-    },
-};
-
 export const pathListChunk = {
     path: "/iplant/home/sriram/ht-limit",
     user: "sriram",
